refactor: file output types
